Fixed an error condition in bitlify: if a string is passed to it and there is a problem, return the string unaltered
Our desired behavior:  Pass a full complete canonical URL of type string to bitlify in a template and get a shortened URL.

Before this checkin, this works EXCEPT in the case of http://localhost[:xxxx]/.  The main bit.ly site won’t shorten it (returns an INVALID_URI error).  This fails which causes an exception it django-bitly the throw an error.  The templatetag bitly wrapper catches the error and tries to call .get_absolute_url() on the passed in object.  But the passed in object is a string, and does not have this parameter.  So the attempt to call the invalid method generates another execption, whtich causes the entire template to fail rendering.

This is a small patch fix for our purposes.  AFAICT, there is a quirk in the framework where passed in strings (instead of objects) aren't handled consistently.  This would ideally be fixed on a deeper level.  Add to "Someday Maybe" list.